bug fixing and improvments
bug fixing and improvements in binaryformatter minifier

I have also saved it using VS 2019

The internal test app has been renamed to be more unique as we use it to check whether command was executed
