+++ TestConsoleApp/Program.cs
@@ -1,12 +1,14 @@
 ﻿using System;
 
-namespace TestConsoleApp
+namespace TestConsoleApp_YSONET
 {
     class Program
     {
         static void Main(string[] args)
         {
-            Console.WriteLine("This is just for code execution testing");
+            Console.WriteLine("This is just for code execution testing.");
+            Console.WriteLine("Args: ");
+            Console.WriteLine(string.Join(",", args));
             Console.ReadLine();
         }
     }
